Current state
Server8 is a working build that receives API requests containing parameters for Cities,ISP's,Countries and a date range. This then builds the suitable SQL request depending on the endpoint that was used. It Validates the query to ensure it isn't violating certain permissions, then executes the query by sending it to a PostgreSQL database running on my laptop. The previous servers are previous states in the build-up to this model. Server5 is the server in a previous state where it used to simply receive the SQL queries directly as text and then send them onto the database. It can be interacted with using the test4 file which is a simple HTML file that sends a SQL query and then outputs the response.  Server9 will be a touched-up version with rounding and possibly the function to remove outliers in the recorded latencies.